third_party/WebKit/LayoutTests/fast/dom/HTMLTemplateElement/ownerDocument-import-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
HTMLTemplateElement content ownerDocument in HTML imports should differ from import document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS template.ownerDocument is not template.content.ownerDocument
PASS successfullyParsed is true

TEST COMPLETE
